refactor: Remove index file/Update import statment
