+++ src/components/ChanceIcons/index.js
@@ -4,7 +4,7 @@ import { View, StyleSheet } from "react-native";
 import Icon from "react-native-vector-icons/MaterialCommunityIcons";
 import PropTypes from "prop-types";
 
-import { iconNames, iconSizes, colors } from "../../common";
+import { iconNames, iconSizes, colors } from "../../common/constants";
 
 const MAX_CHANCE = 3;
