+++ src/components/Layouts/Carousel.js
@@ -3,7 +3,7 @@ import React from "react";
 import { FlatList, StyleSheet, View } from "react-native";
 import PropTypes from "prop-types";
 
-import { game } from "../../common";
+import { game } from "../../common/property";
 
 const Carousel = ({ content, renderItem, size }) => {
   return (
