+++ src/components/Modals/FailModalContainer.js
@@ -4,7 +4,7 @@ import { useNavigation } from "@react-navigation/native";
 import PropTypes from "prop-types";
 
 import { ButtonModal, SmallButton } from "..";
-import { navigations, colors } from "../../common";
+import { navigations, colors } from "../../common/constants";
 
 const FailModalContainer = ({ onRetryCurrentStage }) => {
   const [modalVisible, setModalVisible] = useState(true);
