+++ src/components/Modals/NextStageModalContainer.js
@@ -3,7 +3,7 @@ import React, { useState } from "react";
 import PropTypes from "prop-types";
 
 import { ButtonModal, SmallButton } from "..";
-import { colors } from "../../common";
+import { colors } from "../../common/constants";
 
 const NextStageModalContainer = ({ onNextStage }) => {
   const [modalVisible, setModalVisible] = useState(true);
