+++ src/features/faceGo/container/FaceGoContainer.js
@@ -1,7 +1,8 @@
 import React, { useState, useEffect, useContext, useMemo } from "react";
 
 import { FaceGo } from "..";
-import { game, time } from "../../../common";
+import { time } from "../../../common/constants";
+import { game } from "../../../common/property";
 import { GameContext } from "../../../context";
 import { createMap } from "../../../utils";
