+++ src/features/faceGo/container/MapContainer.js
@@ -3,7 +3,8 @@ import React, { useEffect } from "react";
 import PropTypes from "prop-types";
 
 import { useCharacter } from "../../../hooks";
-import { map, game } from "../../../common";
+import { map } from "../../../common/constants";
+import { game } from "../../../common/property";
 import { Map } from "../presenter";
 
 const MapContainer = ({
