+++ src/features/faceGo/presenter/MapCell.js
@@ -4,7 +4,7 @@ import { View, StyleSheet } from "react-native";
 import Icon from "react-native-vector-icons/MaterialCommunityIcons";
 import PropTypes from "prop-types";
 
-import { colors } from "../../../common";
+import { colors } from "../../../common/constants";
 
 const MapCell = ({ width, height, bgColor, canMove }) => {
   return (
