+++ src/features/main/presenter/GameGuide.js
@@ -4,7 +4,7 @@ import { StyleSheet, View, Text, ScrollView, Image } from "react-native";
 import Icon from "react-native-vector-icons/MaterialCommunityIcons";
 import PropTypes from "prop-types";
 
-import { colors } from "../../../common";
+import { colors } from "../../../common/constants";
 
 const GameGuide = ({ item, style }) => {
   return (
