+++ src/features/main/presenter/Page.js
@@ -3,7 +3,7 @@ import React from "react";
 import { StyleSheet, View, Text, ScrollView } from "react-native";
 import PropTypes from "prop-types";
 
-import { colors } from "../../../common";
+import { colors } from "../../../common/constants";
 
 const Page = ({ item, style }) => {
   return (
